Moving composite score validation from pullData controller to pull metadata controller.
