+++ app/src/main/java/org/eyeseetea/malariacare/data/database/iomodules/dhis/importer/PullMetadataController.java
@@ -35,6 +35,7 @@ import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.models
         .ProgramStageDataElementExtended;
 import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.models.ProgramStageExtended;
 import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.models.UserAccountExtended;
+import org.eyeseetea.malariacare.data.database.model.CompositeScoreDB;
 import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.data.remote.sdk.PullDhisSDKDataSource;
 import org.eyeseetea.malariacare.data.remote.sdk.SdkQueries;
@@ -269,6 +270,8 @@ public class PullMetadataController implements IPullMetadataController {
             }
         }
 
+        validateCS();
+
         //Fill order and parent scores
         System.out.printf("Building compositeScore relationships...");
         converter.buildScores();
@@ -297,5 +300,34 @@ public class PullMetadataController implements IPullMetadataController {
 
         System.out.printf("Building orgunit hierarchy...");
         return converter.buildOrgUnitHierarchy(assignedOrganisationsUnits);
+    }
+
+    private void validateCS() {
+        Log.d(TAG, "Validate Composite scores");
+        callback.onStep(PullStep.VALIDATE_COMPOSITE_SCORES);
+        List<CompositeScoreDB> compositeScores = CompositeScoreDB.list();
+        for (CompositeScoreDB compositeScore : compositeScores) {
+            if (!compositeScore.hasChildren() && (compositeScore.getQuestions() == null
+                    || compositeScore.getQuestions().size() == 0)) {
+                Log.d(TAG,
+                        "CompositeScoreDB without children and without questions will be removed: "
+                                + compositeScore.toString());
+                compositeScore.delete();
+                continue;
+            }
+            if (compositeScore.getHierarchical_code() == null) {
+                Log.d(TAG, "CompositeScoreDB without hierarchical code will be removed: "
+                        + compositeScore.toString());
+                compositeScore.delete();
+                continue;
+            }
+            if (compositeScore.getComposite_score() == null
+                    && !compositeScore.getHierarchical_code().equals(
+                    CompositeScoreBuilder.ROOT_NODE_CODE)) {
+                Log.d(TAG, "CompositeScoreDB not root and not parent should be fixed: "
+                        + compositeScore.toString());
+                continue;
+            }
+        }
     }
 }
